Add Heka fields to a "meta" section, add payload.
Not all of the payload is pulled out into separate fields, so the
base object should start with the protobuf "payload" field. See
https://github.com/mozilla-services/data-pipeline/blob/master/heka/sandbox/decoders/extract_telemetry_dimensions.lua#L168-L169
For what is currently pulled out into separate fields.

Also, introduce a "meta" section to hold the top-level message fields,
such as Timestamp and Type, as well as the fields that are not part
of the original incoming payload (such as geoCountry, sampleId, DNT).