Use development version of happybase
The happybase library has recently gained the ability to scan HBase in
reverse order. Unfortunately the change is only available in the master
branch.

This patch is temporary and will be reverted once a new version of
happybase is available on pypi.

Note that the package has to be installed with pip using
the `--process-dependency-links` flag; without it this patch has
no effect.