Bug 1318326 - Fix `get_pings_properties` for keyedHistograms (#150)
* Bug 1318326 - Fix get_pings_properties for keyedHistograms

This is caused by a KeyError exception that prematurely skipped over the
values of the histograms. The KeyError is triggered when a child process
contains histograms of interest.

* Bug 1318326 - Add failing keyedHistogram test cases for get_pings_properties

* Assert the simple expected behavior case of a keyedHistogram
* Assert the case when children processes also contain keyedHistograms
* Assert behavior when `with_process=True`

* Use a safe dict as ping cursor

The PingCursor allows for legible traversal of nested dictionaries by
defaulting to an empty instance of a subclassed dictionary. The cursor
will also cast all existing sub-dictionaries to this new type.