Bug 1380438 - Make scalar names case-insensitive (#170)
* Make scalar names case-insensitive

We are guaranteed that no two scalars in scalars.yaml will share
a case-insensitive name. Currently, the Scalar class fails to
return a scalar when given a proper scalar name that is not
cased correctly. This change matches any case.

* Add Scalar exception when no definition found

Previously, when a definition wasn't found for a Scalar the
exception was "Missing required elements in definition", which while
true does not indicate that the scalar was not present in the
definition file. This error is a KeyError, and indicates the
metric is missing.